Updated send table processing
Added in the logic needed to process fields and field types in a fairly effecient manor... cross validated results with other CS2 parsers, as they should be parsing this section the same since it's game agnostic.

Everything seems to line up, and this sets up everything needed to properly push on to the serializers, since we have the field definitions now.

Right now the context is capturing a list of all these things, but those should be consisdered temporary as this data is discardable after it is done with, only needed for debugging.