Updated to start reading entities
Hit an issue when reading entities that requires string tables to be implemented to load the baseline data properly which is required for default values of our classes.

Basic parsing is now supported for entity packets, but will have to be paused for now while string table implementation is ported over.