Somewhat working first draft of entity parsing.
Using a lot of burrowed code from the other C# parser, but attempting to port it over without having to define crazy deserializers for higher order types.

Will create a type generator system that can handle entities on the fly and make it much easier to keep everything in date (especially with how often mobas update)

just outputting data to console for now, once it is stable can move on to cleaning this up and optmizing it.