BROKEN COMMIT (WIP)
This is a commit of the code so far, but is incredibly broken and does not actually complete the parsing process.

Publishing just to share progress and let people see the general layout that is being formed.

In the coming days the entity system will be completed, and at which point fully parsing should be nearly complete.

PLEASE DO NOT CONSIDER THIS A WORKING PROJECT YET. JUST REFERENCE THE STUFF IM DOING.