Fixed parse to nearly complete first entity
The first parser testing has shown itself to be stable, however it is not fully parsing the correct values for some fields.

This is known by the values being wildy off from expectations for a baseline of the entity, as well as comparing to other parsers by hacking into their deserializers to extract comparison data reads.

These will be smoothed out shortly but this is stable enough people can look at it to get an idea of the final layout now.