Improved entity parsing stability
Entity parsing has solidified into it's design and shown itself to be working properly, and somewhat flexible enough to properly handle all cases that might come up.

There are still a few unhandled situations and types that I need to add handlers for, wether that is adding defined objects to handle (if they are common and static types or discarded data) or figuring out a way to support them dynamically

Should be able to parse through a few non animation entities, and has been configured so that it will isntead of trying to directly read the packet just iterate through every single class we got an instance baseline for, and print out an instantiated instance for it, catching errors and going onto the next to see what failures are happening.