Fixed an error caused by serializer versioning
Since serializers have a version, and fields will request a specific version we need to make sure to respect that or we will not actually be able to parse out most entities.

This was unfortunately overlooked as the default field that was retrieved for CWorld which was the one being tested was correct by chance which lead to wrong assumptions (a lesson here about avoiding assumptions I guess)

Added in a new search by version for field deserializers so that the correct decoders will be applied when grabbing out the data during object activation.

Created a new test to see how good deserializer handling is and reached down to 12 pending exception types (unique field failures) that will need to be handled.