Very slow implementation of entity update events
Added in a very very quick and dirty solution to propagating the entity update events to the consumer interface, but it is super slow and not long term practical. The interface will likely change drastically so that people will subscribe to very specific things they want to see instead of entire object snapshots.