Rewrote entity parsing system to use concrete types
The old system was returning dictionaries, but this rewrite focuses on transforming the data into a more concrete type that is useful for actually turning into real data.

Updated so that the send table can be dumped to produce class definitions which will be used to generate mapped types.

Once this system is done a generator will be dropped in that will easily produce the concrete types making mapping easier