Wrote code generator for new entity system
Added in a dump functionality to extract mappings using context data in the demo file.

Built a code generator that can parse this dumped data and create concrete types that will be able to get assigned properties using field values for a much faster pushing experience which should help boost performance.

Tested system to the top level types, and need to expand and handle indexed and entity types but after that should be good to go fully.