Added all class definitions
Looks like this is bogging down the build process, I am thinking that going forward I should extract the code for generation to a standalone library that can be called and have it's produced code put into an SDK instead of relying on a roslyn code generator in this case.

Still putting it in for now, as the type provider shouldn't be an issue but is something to be concerned about.