First pass of finalizing the parser interface
Restructured the parser to lean towards an event driven approach which would allow for more adaptability in how the data is processed to make things easier to work with.

Removed some old code and projects breaking the entity system as a replacement will be worked and implemented shortly.

Started laying out interface for data transformations that will make for an easier time reasoning about the flow of data when writing more complex parsing operations